+++ R/bioc.R
@@ -67,6 +67,7 @@
 #' * 2020-03-22 get_matching_bioc_version() is now correct if the current
 #'              R version is not in the builtin mapping.
 #' * 2020-11-21 Update internal map for 3.12.
+#' * 2023-05-08 Add 'books' repo.
 #'
 #' @name bioconductor
 #' @keywords internal
@@ -271,7 +272,9 @@ bioconductor <- local({
       BioCworkflows =
         if (bioc_version >= "3.7") "{mirror}/packages/{bv}/workflows",
       BioCextra     =
-        if (bioc_version <= "3.5") "{mirror}/packages/{bv}/extra"
+        if (bioc_version <= "3.5") "{mirror}/packages/{bv}/extra",
+      BioCbooks     =
+        if (bioc_version >= "3.12") "{mirror}/packages/{bv}/books"
     )
 
     ## It seems that if a repo is not available yet for bioc-devel,