+++ NEWS.md
@@ -1,3 +1,5 @@
+# pkgcache (development version)
+
 # pkgcache 2.1.0
 
 * pkgcache now supports binary packages on x86_64 macOS and R 4.3.0 and